content/media/webaudio/test/test_bug827541.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Tell the cycle collector about the audio contexts owned by nsGlobalWindow</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
  SpecialPowers.setBoolPref("media.webaudio.enabled", true);

  var iframe = document.createElementNS("http://www.w3.org/1999/xhtml", "iframe");
  document.body.appendChild(iframe);
  var frameWin = iframe.contentWindow;
  frameWin.AudioContext();
  document.body.removeChild(iframe);
  frameWin.AudioContext();

  ok(true, "This test should not leak");
  SpecialPowers.clearUserPref("media.webaudio.enabled");
</script>
</pre>
</body>
</html>
